pyproject.toml
[tool.poetry]
name = "Unified Shopping List"
version = "0.3.5"
description = "Open source app to connect your shopping lists in real time"
authors = ["Michael Genson <71845777+michael-genson@users.noreply.github.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.81.0"
pydantic = "^1.10.1"
pyhumps = "^3.7.3"
uvicorn = "^0.18.3"
mangum = "^0.15.1"
dynamodb-json = "^1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
cachetools = "^5.2.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
requests = "^2.28.1"
todoist-api-python = "^2.0.2"
python-dateutil = "^2.8.2"
pytz = "^2022.7"
moto = {extras = ["dynamodb"], version = "^4.1.6"}

[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["essential", "secretsmanager"], version = "^1.24.65"}
boto3 = "^1.24.65"
types-boto3 = "^1.0.2"
types-passlib = "^1.7.7"
types-python-jose = "^3.3.4"
types-cachetools = "^5.2.1"
types-requests = "^2.28.11.5"
fuzzywuzzy-stubs = "^0.0.1"
black = "^22.6.0"
pylint = "^2.14.5"
mypy = "^0.991"
bumpversion = "^0.6.0"
types-python-dateutil = "^2.8.19.5"
types-pytz = "^2022.7.0.0"
pytest = "^7.2.2"
freezegun = "^1.2.2"
pytest-cov = "^4.0.0"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
